refactor: move editor out from default-tiptap
