+++ app/components/editor/styles.scss
